still on still

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -1,7 +1,7 @@
 #include "lists.h"
 #include <stdint.h>
 
-int print_listformat(const listint_t *head);
+size_t print_listformat(const listint_t *head);
 
 /**
  * print_listint_safe - prints lists with loops
@@ -57,7 +57,7 @@ size_t print_listint_safe(const listint_t *head)
  *
  * Return: no of elements in list
 */
-int print_listformat(const listint_t *head)
+size_t print_listformat(const listint_t *head)
 {
 	const listint_t *temp = head;
 	size_t i = 0;
